Added the snackbar functions to views, links to the planner and reset to forms
Added the snack bar functions to views, making them warn when something work (and shows the link of what was edited/created) and when it doesn't.

Added links to the title of events/dailytasks/listtodo in the planner

Added a reset/cancel function to the base.js and the button to all forms.