Lots of changes - upgrade on the programs Css and upgrade on the sort function to a mult criteria sort function (
Made considerable changes to the Css of the base, in order to improve the visual of the pages.

create a mult criteria sort function that can sort based on up to 3 functions, also made changes to the sort buttons so their html change based on their presence(or lack of), priority and order inside the sort criteria.

Still need to apply this sort function to all pages, atm is limited to test.html